Fixed Warnings in Game Code

+++ TimeActionShooter/Game/GameState.cpp
@@ -2,7 +2,7 @@
 
 //-----------------------------------
 
-GameState::GameState( const Player &player, const std::vector<Enemy> &enemies, const std::vector<Bullet> &bullets, const std::vector<PowerUp> &powerUps, int stage, int timer  )
+GameState::GameState( const Player &player, const std::vector<Enemy> &enemies, const std::vector<Bullet> &bullets, const std::vector<PowerUp> &powerUps, int stage, double timer  )
 {
 	m_player = player;
 	m_enemies = enemies;
